+++ CourtCaseManagement.ApplicationCore/TOs/ResponsibleResponseTO.cs
@@ -5,7 +5,7 @@ namespace CourtCaseManagement.ApplicationCore.TOs
     public class ResponsibleResponseTO
     {
         public Guid? Id { get; set; }
-        public int? Cpf { get; set; }
+        public long? Cpf { get; set; }
         public string Mail { get; set; }
         public string Name { get; set; }
         public string Photograph { get; set; }
